+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/Utilities/SyntacticDocument.cs
@@ -9,27 +9,6 @@ using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis
 {
-    internal readonly record struct DocumentSyntax(DocumentId Id, SourceText Text, SyntaxNode Root)
-    {
-        public SyntaxTree SyntaxTree => Root.SyntaxTree;
-
-        public static async ValueTask<DocumentSyntax> CreateAsync(Document document, CancellationToken cancellationToken)
-        {
-            var text = await document.GetTextAsync(cancellationToken).ConfigureAwait(false);
-            var root = await document.GetRequiredSyntaxRootAsync(cancellationToken).ConfigureAwait(false);
-            return new DocumentSyntax(document.Id, text, root);
-        }
-
-#if !CODE_STYLE
-        public static DocumentSyntax CreateSynchronously(Document document, CancellationToken cancellationToken)
-        {
-            var text = document.GetTextSynchronously(cancellationToken);
-            var root = document.GetRequiredSyntaxRootSynchronously(cancellationToken);
-            return new DocumentSyntax(document.Id, text, root);
-        }
-#endif
-    }
-
     internal class SyntacticDocument
     {
         public readonly Document Document;
